refactor: Separa a aplicação em módulos lógicos
- Desmembra o monolítico `main.py` em múltiplos arquivos com responsabilidades claras:
  - `app.py`: Ponto de entrada e interação com o usuário.
  - `graph.py`: Lógica do LangGraph (estado, nós, construção do grafo).
  - `config.py`: Configurações globais (nome do modelo, temperaturas).
  - `prompts.py`: Todos os prompts de especialista.
- Esta refatoração melhora a manutenibilidade, legibilidade e escalabilidade do projeto, alinhando-o com as melhores práticas de arquitetura de software.
