+++ config.py
@@ -1,0 +1,3 @@
+NOME_MODELO = "gemini-1.5-flash"
+CONFIG_GERACAO_ANALISE = {"temperature": 0.2}
+CONFIG_GERACAO_RELATORIO = {"temperature": 0.3}
